refs #3464 : bugfixing displays with a different ration than 1 or 2.
git-svn-id: svn://develop.studip.de/studip/trunk@27284 c5b85792-5220-0410-ad23-ed5b6b8ad999

+++ lib/phplib_local.inc.php
@@ -590,7 +590,7 @@ class Seminar_Auth extends Auth {
 
         $this->auth["uname"] = Request::get('loginname');   // This provides access for "loginform.ihtml"
         $this->auth["jscript"] = Request::get('resolution') != "";
-        $this->auth['devicePixelRatio'] = Request::int('device_pixel_ratio');
+        $this->auth['devicePixelRatio'] = Request::float('device_pixel_ratio');
 
         $check_auth = StudipAuthAbstract::CheckAuthentication(Request::get('loginname'),Request::get('password'),$this->auth['jscript']);
